pyproject.toml
[tool.black]
line-length = 110
target-version = ['py37']
skip-string-normalization = true

[tool.poetry]
name = "linked-list"
version = "0.1.0"
description = "Python implementations of various kinds of linked list data structures"
authors = ["Mike Lane <mikelane@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"

[tool.poetry.dev-dependencies]
black = "^19.3b0"
pytest = "^4.0"
pytest-factoryboy = "^2.0.2"
pytest-lambda = "^0.1"
mypy = "^0.701.0"

[build-system]
requires = ["poetry>=0.12.15"]
build-backend = "poetry.masonry.api"
